refactor engine_monitor to be in common
